// src/app/api/fn/partners/partners-customers-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedCustomerList } from '../../models/paginated-customer-list';

export interface PartnersCustomersList$Params {

/**
 * true: active only; false: inactive only.
 */
  active?: 'false' | 'true';

/**
 * yes: allowed credit (on hold included); no: cash only; hold: on credit hold.
 */
  credit?: 'hold' | 'no' | 'yes';

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Price tier.
 */
  price_tier?: 'RETAIL' | 'WHOLESALE';

/**
 * Matches codes, names and numbers.
 */
  search?: string;
}

export function partnersCustomersList(http: HttpClient, rootUrl: string, params?: PartnersCustomersList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedCustomerList>> {
  const rb = new RequestBuilder(rootUrl, partnersCustomersList.PATH, 'get');
  if (params) {
    rb.query('active', params.active, {});
    rb.query('credit', params.credit, {});
    rb.query('page', params.page, {});
    rb.query('price_tier', params.price_tier, {});
    rb.query('search', params.search, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedCustomerList>;
    })
  );
}

partnersCustomersList.PATH = '/api/partners/customers/';